requirements.txt
yfinance
pandas
requests
python-dotenv